refactor: shuttle file structor
