+++ public/locales/zh-CN.json
@@ -10,5 +10,6 @@
   "searchTokenPlaceholder": "搜索代币名称或者合约地址",
   "selectToken": "选择代币",
   "commonTokens": "常用代币",
-  "tokenList": "代币列表"
+  "tokenList": "代币列表",
+  "next": "下一步"
 }
