+++ src/assets/svg/index.js
@@ -6,5 +6,6 @@ export {ReactComponent as Loading} from './loading.svg'
 export {ReactComponent as Search} from './search.svg'
 export {ReactComponent as ArrowLeft} from './arrow-left.svg'
 export {ReactComponent as PlusWithBg} from './bg-plus.svg'
+export {ReactComponent as ChangeWithBg} from './bg-change.svg'
 export {ReactComponent as ArrowRightWithBg} from './bg-arrow-right.svg'
 export {ReactComponent as Plus} from './plus.svg'
