+++ src/pages/components/index.js
@@ -1,2 +1,3 @@
 export {default as TokenNameAndIcon} from './TokenNameAndIcon'
 export {default as TransactionReceiptionModal} from './TransactionReceiptionModal'
+export {default as ChainItem} from './ChainItem'
